# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Server pets",
        "title": "Swagger Example API",
        "contact": {},
        "license": {},
        "version": "0.0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/autoCode/createTemp": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SysApi"
                ],
                "summary": "自动代码模板",
                "parameters": [
                    {
                        "description": "创建自动代码",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sys_model.AutoCodeStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"创建成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "sys_model.AutoCodeStruct": {
            "type": "object",
            "properties": {
                "abbreviation": {
                    "type": "string"
                },
                "autoCreateApiToSql": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/sys_model.Field"
                    }
                },
                "packageName": {
                    "type": "string"
                },
                "structName": {
                    "type": "string"
                },
                "tableName": {
                    "type": "string"
                }
            }
        },
...
        "sys_model.Field": {
            "type": "object",
            "properties": {
                "columnName": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "dataType": {
                    "type": "string"
                },
                "dataTypeLong": {
                    "type": "string"
                },
                "dictType": {
                    "type": "string"
                },
                "fieldDesc": {
                    "type": "string"
                },
                "fieldJson": {
                    "type": "string"
                },
                "fieldName": {
                    "type": "string"
                },
                "fieldSearchType": {
                    "type": "string"
                },
                "fieldType": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-token",
            "in": "header"
        }
    }
}